requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
google==3.0.0
idna==3.7
mysql-connector-python==8.4.0
requests==2.32.2
schedule==1.2.1
soupsieve==2.5
urllib3==2.2.1
